pmm ready

+++ src/arch/i386/mm/pmm.h
@@ -18,13 +18,13 @@
 #define PMM_PAGE_SIZE 0x1000      // 页面大小 4KB
 #define PAGE_MAX_SIZE (PMM_MAX_SIZE / PMM_PAGE_SIZE)
 
-extern uint64_t kernel_start[];
-extern uint64_t kernel_end[];
+extern uint32_t kernel_start[];
+extern uint32_t kernel_end[];
 
 // 初始化内存管理
 void pmm_init(multiboot_tag_t * tag);
 
-uint64_t pmm_alloc_page(void);
+uint32_t pmm_alloc_page(void);
 
-void pmm_free_page(uint64_t page);
+void pmm_free_page(uint32_t page);
 #endif
